+++ imod/tests/test_mf6/test_mf6_simulation.py
@@ -13,10 +13,13 @@ import numpy as np
 import pandas as pd
 import pytest
 import rasterio
+import tomli
+import tomli_w
 import xarray as xr
 import xugrid as xu
 
 import imod
+from imod.logging import LoggerType, LogLevel
 from imod.mf6 import LayeredWell, Well
 from imod.mf6.model import Modflow6Model
 from imod.mf6.multimodel.modelsplitter import PartitionInfo
@@ -73,6 +76,89 @@ def test_circle_roundtrip(circle_model, tmpdir_factory):
 
 
 @pytest.mark.usefixtures("twri_model")
+def test_dump_version_number__version_written(twri_model, tmpdir_factory):
+    # Arrange
+    tmp_path = tmpdir_factory.mktemp("twri")
+    # Act
+    twri_model.dump(tmp_path)
+    # Assert
+    toml_path = tmp_path / f"{twri_model.name}.toml"
+    with open(toml_path, "rb") as f:
+        toml_content = tomli.load(f)
+    assert toml_content["version"]["imod-python"] == imod.__version__
+
+
+@pytest.mark.usefixtures("twri_model")
+def test_from_file_version_logged__version_in_dumped(twri_model, tmpdir_factory):
+    """
+    Tested if a warning is thrown when there is a mismatch between version
+    numbers of saved model and current iMOD Python version.
+    """
+    # Arrange
+    tmp_path = tmpdir_factory.mktemp("twri")
+    twri_model.dump(tmp_path)
+    toml_path = tmp_path / f"{twri_model.name}.toml"
+    with open(toml_path, "rb") as f:
+        toml_content = tomli.load(f)
+    toml_content["version"]["imod-python"] = "0.0.0"
+
+    toml_path_adapted = tmp_path / f"{twri_model.name}_adapted.toml"
+    with open(toml_path_adapted, "wb") as f:
+        tomli_w.dump(toml_content, f)
+    # Act
+    logfile_path = tmp_path / "logfile.txt"
+    with open(logfile_path, "w") as sys.stdout:
+        imod.logging.configure(
+            LoggerType.PYTHON,
+            log_level=LogLevel.DEBUG,
+            add_default_file_handler=False,
+            add_default_stream_handler=True,
+        )
+        imod.mf6.Modflow6Simulation.from_file(toml_path_adapted)
+
+    # Assert
+    with open(logfile_path, "r") as log_file:
+        log = log_file.read()
+        assert f"iMOD Python version in current environment: {imod.__version__}" in log
+        assert "iMOD Python version in dumped simulation: 0.0.0" in log
+
+
+@pytest.mark.usefixtures("twri_model")
+def test_from_file_version_logged__no_version_in_dumped(twri_model, tmpdir_factory):
+    """
+    Tested if a warning is thrown when there is a mismatch between version
+    numbers of saved model and current iMOD Python version.
+    """
+    # Arrange
+    tmp_path = tmpdir_factory.mktemp("twri")
+    twri_model.dump(tmp_path)
+    toml_path = tmp_path / f"{twri_model.name}.toml"
+    with open(toml_path, "rb") as f:
+        toml_content = tomli.load(f)
+    toml_content.pop("version")
+
+    toml_path_adapted = tmp_path / f"{twri_model.name}_adapted.toml"
+    with open(toml_path_adapted, "wb") as f:
+        tomli_w.dump(toml_content, f)
+    # Act
+    logfile_path = tmp_path / "logfile.txt"
+    with open(logfile_path, "w") as sys.stdout:
+        imod.logging.configure(
+            LoggerType.PYTHON,
+            log_level=LogLevel.DEBUG,
+            add_default_file_handler=False,
+            add_default_stream_handler=True,
+        )
+        imod.mf6.Modflow6Simulation.from_file(toml_path_adapted)
+
+    # Assert
+    with open(logfile_path, "r") as log_file:
+        log = log_file.read()
+        assert f"iMOD Python version in current environment: {imod.__version__}" in log
+        assert "No iMOD Python version information found in dumped simulation." in log
+
+
+@pytest.mark.usefixtures("twri_model")
 def test_twri_gdal(twri_model, tmpdir_factory):
     """
     Test of dumping a structured model with a CRS results in the necessary
@@ -107,7 +193,6 @@ def test_twri_disv_mdal_compliant_semi_roundtrip(twri_disv_model, tmpdir_factory
 
 
 @pytest.mark.usefixtures("circle_model")
-@pytest.mark.skipif(sys.version_info < (3, 7), reason="capture_output added in 3.7")
 def test_simulation_open_head(circle_model, tmp_path):
     simulation = circle_model
 
@@ -136,7 +221,6 @@ def test_simulation_open_head(circle_model, tmp_path):
 
 
 @pytest.mark.usefixtures("circle_model")
-@pytest.mark.skipif(sys.version_info < (3, 7), reason="capture_output added in 3.7")
 def test_simulation_open_head_relative_path(circle_model, tmp_path):
     simulation = circle_model
 
@@ -157,7 +241,6 @@ def test_simulation_open_head_relative_path(circle_model, tmp_path):
 
 
 @pytest.mark.usefixtures("circle_model")
-@pytest.mark.skipif(sys.version_info < (3, 7), reason="capture_output added in 3.7")
 def test_simulation_open_flow_budget(circle_model, tmp_path):
     simulation = circle_model
 
@@ -197,7 +280,6 @@ def test_write_circle_model_twice(circle_model, tmp_path):
 
 
 @pytest.mark.usefixtures("circle_model")
-@pytest.mark.skipif(sys.version_info < (3, 7), reason="capture_output added in 3.7")
 def test_simulation_open_concentration_fail(circle_model, tmp_path):
     """No transport model is assigned, so should throw error when opening concentrations"""
     simulation = circle_model